SharpFPS: Implement "Single Point Frame Pacing"
Single point frame pacing is a modification of the original SharpFPS that allows you to limit the framerate by basically placing a call to SharpFPS at a single place within the recurring logic loop, as opposed to wrapping the start and end of recurring loop with StartFrame and EndFrame.

This is to allow for frame pacing in cases where there is external code outside of the user's control, where a StartFrame or EndFrame cannot be used.